Fixing the names and the creation of the first migration

+++ src/main/java/com/codigo/monefy/transactions/persistence/entity/Account.java
@@ -5,6 +5,8 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "accounts")
 @Getter
@@ -23,15 +25,15 @@ public class Account {
     private AccountType accountType;
     private String name;
     private String description;
-    private Double initialBalance;
-    private Double currentBalance;
+    private BigDecimal initialBalance;
+    private BigDecimal currentBalance;
     private boolean isActive;
     private String bankName;
     private String accountNumber;
 
     public Account() {}
 
-    public Account(Integer id, User user, AccountType accountType, String name, String description, Double initialBalance, Double currentBalance, boolean isActive, String bankName, String accountNumber) {
+    public Account(Integer id, User user, AccountType accountType, String name, String description, BigDecimal initialBalance, BigDecimal currentBalance, boolean isActive, String bankName, String accountNumber) {
         this.id = id;
         this.user = user;
         this.accountType = accountType;
