+++ src/main/java/com/codigo/monefy/transactions/persistence/entity/FinancialGoal.java
@@ -5,6 +5,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -23,8 +24,8 @@ public class FinancialGoal {
 
     private String name;
     private String description;
-    private Double targetAmount;
-    private Double currentAmount;
+    private BigDecimal targetAmount;
+    private BigDecimal currentAmount;
     private LocalDate targetDate;
     private String goalType;
     private boolean isAchieved;
@@ -32,7 +33,7 @@ public class FinancialGoal {
 
     public FinancialGoal() {}
 
-    public FinancialGoal(Integer id, User user, String name, String description, Double targetAmount, Double currentAmount, LocalDate targetDate, String goalType, boolean isAchieved, boolean isActive) {
+    public FinancialGoal(Integer id, User user, String name, String description, BigDecimal targetAmount, BigDecimal currentAmount, LocalDate targetDate, String goalType, boolean isAchieved, boolean isActive) {
         this.id = id;
         this.user = user;
         this.name = name;
