+++ src/main/java/com/codigo/monefy/transactions/persistence/entity/Transaction.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -23,7 +24,7 @@ public class Transaction {
     @JoinColumn(name = "category_id", nullable = false)
     private Category category;
 
-    private Double amount;
+    private BigDecimal amount;
     private String description;
     private LocalDateTime transactionDate;
     private String transactionType; // e.g., "income" or "expense"
@@ -32,7 +33,7 @@ public class Transaction {
 
     public Transaction() {}
 
-    public Transaction(Integer id, Account account, Category category, Double amount, String description, LocalDateTime transactionDate, String transactionType, String notes, String location) {
+    public Transaction(Integer id, Account account, Category category, BigDecimal amount, String description, LocalDateTime transactionDate, String transactionType, String notes, String location) {
         this.id = id;
         this.account = account;
         this.category = category;
